+++ replaying-share-kotlin/src/main/kotlin/com/jakewharton/rx3/ReplayingShare.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.jakewharton.rx
+package com.jakewharton.rx3
 
 import io.reactivex.rxjava3.core.Flowable
 import io.reactivex.rxjava3.core.Observable
