+++ replaying-share/src/test/java/com/jakewharton/rx3/ReplayingShareObservableTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.jakewharton.rx;
+package com.jakewharton.rx3;
 
 import io.reactivex.rxjava3.core.Observable;
 import io.reactivex.rxjava3.disposables.Disposable;
